# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CoopData API",
    "description": "Cooperative Data Management Platform API with RBAC hierarchy (Ministry \u2192 Federation \u2192 Apex \u2192 Cooperative)",
    "contact": {
      "name": "CoopData Team",
      "email": "support@coopdata.org"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/cooperative/financial-statements/{id}": {
      "get": {
        "tags": [
          "Cooperative"
        ],
        "operationId": "get_financial_statement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Financial statement ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Financial statement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialStatementResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/v1/cooperative/financial-statements/{id}/line-items": {
      "get": {
        "tags": [
          "Cooperative"
        ],
        "operationId": "list_line_items",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Financial statement ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Line items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LineItemResponse"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Cooperative"
        ],
        "operationId": "update_line_items",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Financial statement ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemBulkUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Line items updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LineItemResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FinancialStatementResponse": {
        "type": "object",
        "required": [
          "id",
          "submission_id",
          "cooperative_id",
          "reporting_year",
          "accounting_year",
          "currency",
          "is_validated",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "accounting_year": {
            "type": "string"
          },
          "cooperative_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "currency": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_validated": {
            "type": "boolean"
          },
          "reporting_year": {
            "type": "integer",
            "format": "int32"
          },
          "submission_id": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "validation_errors": {
            "nullable": true
          }
        }
      },
...
      "LineItemBulkUpdateRequest": {
        "type": "object",
        "required": [
          "updates"
        ],
        "properties": {
          "updates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LineItemUpdateRequest"
            }
          }
        }
      },
      "LineItemResponse": {
        "type": "object",
        "required": [
          "id",
          "financial_statement_id",
          "account_name",
          "account_category",
          "account_subcategory",
          "month",
          "ai_flagged",
          "manually_edited",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "account_category": {
            "type": "string"
          },
          "account_code": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "account_name": {
            "type": "string"
          },
          "account_subcategory": {
            "type": "string"
          },
          "ai_confidence": {
            "type": "number",
            "format": "double",
            "description": "AI extraction confidence (0.0\u20131.0)",
            "nullable": true
          },
          "ai_flagged": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "financial_statement_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "manually_edited": {
            "type": "boolean"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "raw_label": {
            "type": "string",
            "description": "Original text label from the uploaded document",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "type": "number",
            "format": "double",
            "description": "Monetary value",
            "nullable": true
          }
        }
      },
      "LineItemUpdateRequest": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "account_code": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "value": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
...
    }
  }
}